+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - global utils functions
 - hot reload for dev
 - emote command to fetch all emotes for channel
+- save emotegame results as stats
 ### Changed
 
 - remove channel from Response