+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.2.2](https://github.com/helltf/helltfbot-v2/releases/tag/v1.2.2) - 2022-07-01
 
 ### Added
 
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fix
 
 - no longer sending letters after word has been guessed
+
 ## [1.2.1](https://github.com/helltf/helltfbot-v2/releases/tag/v1.2.1) - 2022-06-08
 
 ### Added